iverilog: Allow tool path to have spaces in it
When invoking the various subcommands make sure the path to the executable
is quoted. This allows it to contain spaces.

On Windows quoting the command passed to `system()` or `popen()` is not
enough.  The default behavior under windows will strip quotes from the
beginning and end of the command string under certain conditions. For this
reason implement `system()` and `popen()` equivalents that do not suffer
from this.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>